requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

sqlalchemy>=2.0.0
psycopg2-binary==2.9.9
alembic>=1.13.0

python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

groq
PyGithub>=2.1.1
email-validator>=2.0.0